Add RosettaPinElement tests!
